pyproject.toml
[tool.poetry]
name = "peskyoriginalconsulting"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-binance = "^1.0.12"
xlsxwriter = "^1.4.3"
"pushbullet.py" = "^0.12.0"
"currency.converter" = "^0.5.5"
CurrencyConverter = "^0.16.1"
flask = "^2.0.1"
websocket-client = "^1.1.0"
replit = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
